+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGE LOG
 
 ## 1.1.0
-* [Enhancement] .NET 6 Support for Android, iOS, and MacCatalyst
+* [Enhancement] .NET 6 Support for Android, iOS, and MacCatalyst initial work - still some bugs with MAUI assets
+* [Enhancement] Option to set where JSON config is optional or not
 * [Fix] Detect if Android asset is available
 
 ## 1.0